requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
joblib
gdown
tensorflow
pillow